Reorganize project structure for improved maintainability
- Move configuration files to dedicated 'config' directory
- Relocate database migration files to 'drizzle' folder
- Create 'src' directory for main application code
  - Separate API, database, routes, and schemas into subdirectories
- Add 'scripts' folder for utility scripts
- Move tests to dedicated 'tests' directory
- Update root-level files (package.json, tsconfig.json, etc.)

This restructuring improves code organization, making the project
more maintainable and easier to navigate for developers.
