--- api-types.ts
@@ -1,6 +1,0 @@
-import { z, type ZodError } from 'zod';
-
-export interface ErrorResponse {
-  error: string;
-  details?: ZodError['errors'];
-}
