+++ scripts/register-commands.ts
