+++ src/schemas/errors.ts
