+++ tsconfig.json
@@ -7,7 +7,7 @@
       "esModuleInterop": true,
       "skipLibCheck": true,
       "forceConsistentCasingInFileNames": true,
-      "baseUrl": ".",
+      "baseUrl": "src",
       "paths": {
         "@/*": ["./*"]
       }
